Add controller methods for crud

+++ sklep/app/controllers/UserController.scala
@@ -10,15 +10,15 @@ class UserController @Inject()(cc: ControllerComponents) extends AbstractControl
     Ok(views.html.index("This is getAllUsers view"))
   }
 
-  def getUserById(id: Integer) = Action {
+  def getUserById(id: Long) = Action {
     Ok(views.html.index("This is getUserById view"))
   }
 
-  def getUserOrders(id: Integer) = Action {
+  def getUserOrders(id: Long) = Action {
     Ok(views.html.index("This is getUserOrders view"))
   }
 
-  def getUserDetails(id: Integer) = Action {
+  def getUserDetails(id: Long) = Action {
     Ok(views.html.index("This is getUserDetails view"))
   }
 
@@ -26,7 +26,7 @@ class UserController @Inject()(cc: ControllerComponents) extends AbstractControl
     Ok(views.html.index("This is addUser action"))
   }
 
-  def updateUser(id: Integer) = Action {
+  def updateUser(id: Long) = Action {
     Ok(views.html.index("This is updateUser action"))
   }
 }
